Moved the notification permissions to the login screen as its need for when the user forgets their password
