+++ presentation/src/main/java/me/androidbox/presentation/home/HomeState.kt
@@ -4,8 +4,7 @@ data class HomeState(
     val homeItems: List<HomeItems> = emptyList(),
     val isLoading: Boolean = false,
     val isSuccessLogout: Boolean = false,
-    val showShowDialog: Boolean = false,
-    val showNotificationRationale: Boolean = false
+    val showShowDialog: Boolean = false
 )
